+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 - Fixed uncraftable potion crash
 - Fixed chat interference with hydration ui
 - Fixed bamboo pump crash
+- Fixed puddles compat
 ### Changed:
 -